Update AIS tutorial and create mkdocs for tutorials
