+++ requirements.txt
@@ -11,3 +11,4 @@ numpy<2.0
 pyarrow
 pydantic
 exactextract
+arlas_cli
